requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0

# API & Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
boto3>=1.28.0
redis>=5.0.0

# ML Ops & Model Management
mlflow>=2.8.0
wandb>=0.16.0
dvc>=3.30.0

# LLM Integration
openai>=1.3.0
anthropic>=0.7.0
transformers>=4.35.0

# Data Processing
pytz>=2023.3
python-dateutil>=2.8.2
requests>=2.31.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
